+++ src/Stormancer.Plugins/GameFinder/Stormancer.Server.Plugins.GameFinder/Changelog.rst
@@ -9,6 +9,11 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+Unreleased
+-----------
+Added
+*****
+- Added party gamefinder
 6.1.0.13
 ----------
 Fixed